In the CoreNLPManipulator.java created 3 functions: testStanfordEnglish(): to test the Stanford Core NLP library on an English sentence testStanfordGerman(): to test the Stanford Core NLP library on a German sentence testStanfordItalian(): to test Tint, the library for Italian built upon the Stanford Core NLP, on an Italian sentence
The functions for English and German work, the function for Italian does not due to dependency problems (I have to continue working on the pom.xml)

Introduced changes to the pom.xml trying to intergate Tint (for Italian), but have not succeeded yet.

Copied the file StanfordCoreNLP-german.properties into the src/main/resources folder from the ./m2/repository/edu/stanford/nlp/stanford-corenlp/3.9.2 so that it is seen during execution.